+++ src/primitive/put/ft_putnlst.c
@@ -6,7 +6,7 @@
 /*   By: rlutt <rlutt@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/20 13:12:39 by rlutt             #+#    #+#             */
-/*   Updated: 2017/03/08 15:00:35 by rlutt            ###   ########.fr       */
+/*   Updated: 2017/04/05 10:42:48 by rlutt            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void		ft_putnlst(t_list *lst)
 	tmp = lst;
 	while (tmp)
 	{
-		ft_putnbr((int)tmp->content);
+		ft_putnbr(tmp->content);
 		ft_putchar('\n');
 		tmp = tmp->next;
 	}
